update article et maraicher

+++ src/main/java/dev/rlaffi/spring/patatedouce/services/AlimentService.java
@@ -1,55 +1,91 @@
 package dev.rlaffi.spring.patatedouce.services;
 
 import dev.rlaffi.spring.patatedouce.entities.Aliment;
+import dev.rlaffi.spring.patatedouce.entities.Maraicher;
 import dev.rlaffi.spring.patatedouce.entities.TypeAliment;
 import dev.rlaffi.spring.patatedouce.entities.Utilisateur;
+import dev.rlaffi.spring.patatedouce.exceptions.PermissionException;
+import dev.rlaffi.spring.patatedouce.exceptions.ResourceNotFoundException;
 import dev.rlaffi.spring.patatedouce.repositories.AlimentRepository;
+import dev.rlaffi.spring.patatedouce.repositories.MaraicherRepository;
 import dev.rlaffi.spring.patatedouce.repositories.TypeAlimentRepository;
-import org.keycloak.admin.client.Keycloak;
-import org.keycloak.admin.client.KeycloakBuilder;
-import org.keycloak.admin.client.resource.RealmResource;
-import org.keycloak.admin.client.resource.UsersResource;
-import org.keycloak.representations.AccessTokenResponse;
-import org.keycloak.representations.idm.CredentialRepresentation;
-import org.keycloak.representations.idm.UserRepresentation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.stereotype.Service;
 
-import javax.ws.rs.core.Response;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
-import java.util.Optional;
+import java.util.NoSuchElementException;
 
 @Service
 public class AlimentService {
 	@Autowired
 	private AlimentRepository alimentRepository;
 	@Autowired
+	private LiaisonUtilisateurService liaisonUtilisateurService;
+	@Autowired
+	private MaraicherRepository maraicherRepository;
+	@Autowired
 	private TypeAlimentRepository typeAlimentRepository;
 	private static final Logger logger = LoggerFactory.getLogger(AlimentService.class);
-	public Aliment get(Integer Id) {
-		return alimentRepository.findById(Id).orElseThrow();
+	public Aliment get(Integer id) throws ResourceNotFoundException {
+		return alimentRepository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException(Aliment.class, id));
 	}
 	public List<Aliment> getAll() {
 		return alimentRepository.findAll();
 	}
-	public void deleteById(Integer alimentId) {
-		alimentRepository.deleteById(alimentId);
+	public void deleteById(Integer alimentId) throws ResourceNotFoundException, PermissionException {
+		try {
+			Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+			Jwt jwt = (Jwt) authentication.getPrincipal();
+			String userId = jwt.getSubject();
+			if (authentication != null && userId != null) {
+				Utilisateur utilisateur = liaisonUtilisateurService.getById(userId);
+				Maraicher maraicher = maraicherRepository.findByEmail(utilisateur.getEmail());
+				Aliment aliment = alimentRepository.findById(alimentId).orElseThrow(() -> new ResourceNotFoundException(Aliment.class, alimentId));
+				if(aliment.getMaraicher() == maraicher){
+					alimentRepository.deleteById(alimentId);
+				}else{
+					throw new PermissionException("supprimer l'aliment.");
+				}
+			}
+		} catch (NoSuchElementException e) {
+			throw new ResourceNotFoundException(Aliment.class, alimentId);
+		}
 	}
 	public Aliment create(Aliment aliment) {
 		if (aliment.getTypeAliment() != null) {
 			TypeAliment typeAliment = aliment.getTypeAliment();
-			TypeAliment typeAlimentFound = typeAlimentRepository.findById(typeAliment.getId()).orElseThrow();
-			if(typeAlimentFound == null){
-				typeAliment = typeAlimentRepository.save(typeAliment);
-				aliment.setTypeAliment(typeAliment);
-			}else{
-				aliment.setTypeAliment(typeAlimentFound);
+			if (typeAliment.getId() != null) {
+				TypeAliment typeAlimentFound = typeAlimentRepository.findById(typeAliment.getId()).orElse(null);
+				if (typeAlimentFound == null) {
+					typeAliment = typeAlimentRepository.save(typeAliment);
+					aliment.setTypeAliment(typeAliment);
+				} else {
+					aliment.setTypeAliment(typeAlimentFound);
+				}
 			}
 		}
+
+		if (aliment.getMaraicher() != null) {
+			Maraicher maraicher = aliment.getMaraicher();
+			if (maraicher.getId() != null) {
+				Maraicher maraicherFound = maraicherRepository.findById(maraicher.getId()).orElse(null);
+				if (maraicherFound == null) {
+					maraicher = maraicherRepository.save(maraicher);
+					aliment.setMaraicher(maraicher);
+				} else {
+					aliment.setMaraicher(maraicherFound);
+				}
+			}
+		}
+
 		return alimentRepository.save(aliment);
 	}
+
 }
